Nuances
Strictly speaking, when the sum of the lift and drag forces on the blade
is oriented so that it matches the forward direction of the kayak, the
path of the blade through the water has a backward component relative to
the direction of the kayak.
While most efficient in terms of forces, this style mandates an
increased stroke rate.